Delimit all commands with \r\n instead of \r
This is becasue we treat all commands as serial line commands.  Thus
we should use proper serial line delimeters to delimit the end of a
command.

As an added bonus this helps us work around a shortcoming in the ESP8266
chip (IPD doesn't terminate lines with \r\n).  This allows our serial
library to always read and handle the line appropriately.